Version 1.0.0 -> 1.0.1
update: don't update db if eHand fail

fix: Util.bluePrintAndScan to handle null

+++ src/main/java/com/mcs/modelsearcher/controller/MainController.java
@@ -113,9 +113,11 @@ public class MainController {
 
                 boolean isDeleteConform = showDeleteConfirmation(item);
                 if (isDeleteConform) {
-                  eCon.deleteFromExcel(insertNo);
-                  eCon.deleteFromDb(insertNo);
-                  afterMakingChange();
+                  int deleteResult = eCon.deleteFromExcel(insertNo);
+                  if (deleteResult == 1) {
+                    eCon.deleteFromDb(insertNo);
+                    afterMakingChange();
+                  }
                 }
               });
